[WEAV-177] 유저 실 도메인 정의

+++ infrastructure/persistence/src/main/kotlin/com/studentcenter/weave/infrastructure/persistence/user/entity/UserSilJpaEntity.kt
@@ -8,7 +8,7 @@ import java.util.*
 class UserSilJpaEntity(
     id: UUID,
     userId: UUID,
-    amount: Int,
+    amount: Long,
 ) {
 
     @Id
@@ -21,7 +21,7 @@ class UserSilJpaEntity(
         private set
 
     @Column(nullable = false)
-    var amount: Int = amount
+    var amount: Long = amount
         private set
 
     companion object {
